+++ packages/sdk/src/index.ts
@@ -1,4 +1,4 @@
-export * from './calls';
+export * from './execute';
 export * from './constants';
 export * from './v1/constants';
 export * from './v1/token';
